requirements.txt
langchain
langchain-anthropic
langchain-community
python-dotenv